Admin price component fix
